# contracts/critical-staff.openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/auth/sessions/current": {
      "get": {
        "operationId": "AuthSessionsController_current",
        "summary": "Đọc Staff hiện tại sau khi token đã đối chiếu DB",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffCurrentSessionEnvelopeDto"
                }
              }
            }
          }
        },
        "tags": [
          "Auth Sessions v1"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "AuthSessionsController_logout",
        "summary": "Thu hồi toàn bộ phiên Staff và tăng token version",
        "parameters": [],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Auth Sessions v1"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/activity-logs": {
      "get": {
        "operationId": "ActivityLogController_findAll",
        "summary": "[Admin] Xem nhật ký hoạt động — lọc theo entity/entityId/userId",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number",
              "example": 20,
              "maximum": 100
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "entityId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "entity",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "action",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedActivityLogEnvelopeDto"
                }
              }
            }
          }
        },
        "tags": [
          "Activity Log"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/simulation/benchmark/run": {
      "post": {
        "operationId": "BenchmarkController_run",
        "summary": "Run one deterministic in-memory routing algorithm benchmark",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunBenchmarkDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BenchmarkRunEnvelopeDto"
                }
              }
            }
          }
        },
        "tags": [
          "Simulation Benchmark"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/simulation/benchmark/compare": {
      "post": {
        "operationId": "BenchmarkController_compare",
        "summary": "Compare routing strategies on one immutable deterministic scenario",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompareBenchmarkDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BenchmarkCompareEnvelopeDto"
                }
              }
            }
          }
        },
        "tags": [
          "Simulation Benchmark"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "HospitalManagement Critical Staff Contract",
    "description": "API cho hệ thống tối ưu luồng khám chữa bệnh",
    "version": "0.1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "BenchmarkCompareEnvelopeDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "statusCode": {
            "type": "integer"
          },
          "requestId": {
            "type": "string"
          },
          "timestamp": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "properties": {
              "scenario": {
                "type": "object",
                "properties": {
                  "patientCount": {
                    "type": "integer"
                  },
                  "workflow": {
                    "type": "string",
                    "enum": [
                      "INDEPENDENT",
                      "SEQUENTIAL",
                      "PARTIAL"
                    ]
                  },
                  "seed": {
                    "type": "integer"
                  },
                  "schemaVersion": {
                    "type": "number",
                    "minimum": 1,
                    "maximum": 1
                  }
                },
                "required": [
                  "patientCount",
                  "workflow",
                  "seed",
                  "schemaVersion"
                ]
              },
              "results": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "algorithm": {
                      "type": "string",
                      "enum": [
                        "SYSTEM",
                        "SHORTEST_QUEUE",
                        "ROUND_ROBIN",
                        "RANDOM",
                        "LEAST_UTILISED"
                      ]
                    },
                    "algorithmLabel": {
                      "type": "string"
                    },
                    "seed": {
                      "type": "integer"
                    },
                    "workflow": {
                      "type": "string",
                      "enum": [
                        "INDEPENDENT",
                        "SEQUENTIAL",
                        "PARTIAL"
                      ]
                    },
                    "patientCount": {
                      "type": "integer"
                    },
                    "simulationTimeMs": {
                      "type": "number"
                    },
                    "metrics": {
                      "type": "object",
                      "properties": {
                        "averageWaitingTimeMs": {
                          "type": "number"
                        },
                        "p95WaitingTimeMs": {
                          "type": "number"
                        },
                        "maxWaitingTimeMs": {
                          "type": "number"
                        },
                        "averageLengthOfStayMs": {
                          "type": "number"
                        },
                        "throughputPerSimHour": {
                          "type": "number"
                        },
                        "averageRoomUtilizationPct": {
                          "type": "number"
                        },
                        "completedPatientCount": {
                          "type": "integer"
                        }
                      },
                      "required": [
                        "averageWaitingTimeMs",
                        "p95WaitingTimeMs",
                        "maxWaitingTimeMs",
                        "averageLengthOfStayMs",
                        "throughputPerSimHour",
                        "averageRoomUtilizationPct",
                        "completedPatientCount"
                      ]
                    },
                    "rooms": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "roomId": {
                            "type": "string"
                          },
                          "serviceId": {
                            "type": "string"
                          },
                          "patientsServed": {
                            "type": "integer"
                          },
                          "busyTimeMs": {
                            "type": "number"
                          },
                          "utilizationPct": {
                            "type": "number"
                          }
                        },
                        "required": [
                          "roomId",
                          "serviceId",
                          "patientsServed",
                          "busyTimeMs",
                          "utilizationPct"
                        ]
                      }
                    },
                    "events": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "simTimeMs": {
                            "type": "number"
                          },
                          "type": {
                            "type": "string",
                            "enum": [
                              "ARRIVED",
                              "STEP_READY",
                              "QUEUED",
                              "SERVICE_STARTED",
                              "SERVICE_COMPLETED",
                              "PATIENT_COMPLETED"
                            ]
                          },
                          "patientId": {
                            "type": "string"
                          },
                          "serviceId": {
                            "type": "string"
                          },
                          "roomId": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "simTimeMs",
                          "type",
                          "patientId"
                        ]
                      }
                    }
                  },
                  "required": [
                    "algorithm",
                    "algorithmLabel",
                    "seed",
                    "workflow",
                    "patientCount",
                    "simulationTimeMs",
                    "metrics",
                    "rooms",
                    "events"
                  ]
                }
              }
            },
            "required": [
              "scenario",
              "results"
            ]
          }
        },
        "required": [
          "success",
          "statusCode",
          "requestId",
          "timestamp",
          "data"
        ]
      },
      "BenchmarkRunEnvelopeDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "statusCode": {
            "type": "integer"
          },
          "requestId": {
            "type": "string"
          },
          "timestamp": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "properties": {
              "algorithm": {
                "type": "string",
                "enum": [
                  "SYSTEM",
                  "SHORTEST_QUEUE",
                  "ROUND_ROBIN",
                  "RANDOM",
                  "LEAST_UTILISED"
                ]
              },
              "algorithmLabel": {
                "type": "string"
              },
              "seed": {
                "type": "integer"
              },
              "workflow": {
                "type": "string",
                "enum": [
                  "INDEPENDENT",
                  "SEQUENTIAL",
                  "PARTIAL"
                ]
              },
              "patientCount": {
                "type": "integer"
              },
              "simulationTimeMs": {
                "type": "number"
              },
              "metrics": {
                "type": "object",
                "properties": {
                  "averageWaitingTimeMs": {
                    "type": "number"
                  },
                  "p95WaitingTimeMs": {
                    "type": "number"
                  },
                  "maxWaitingTimeMs": {
                    "type": "number"
                  },
                  "averageLengthOfStayMs": {
                    "type": "number"
                  },
                  "throughputPerSimHour": {
                    "type": "number"
                  },
                  "averageRoomUtilizationPct": {
                    "type": "number"
                  },
                  "completedPatientCount": {
                    "type": "integer"
                  }
                },
                "required": [
                  "averageWaitingTimeMs",
                  "p95WaitingTimeMs",
                  "maxWaitingTimeMs",
                  "averageLengthOfStayMs",
                  "throughputPerSimHour",
                  "averageRoomUtilizationPct",
                  "completedPatientCount"
                ]
              },
              "rooms": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "roomId": {
                      "type": "string"
                    },
                    "serviceId": {
                      "type": "string"
                    },
                    "patientsServed": {
                      "type": "integer"
                    },
                    "busyTimeMs": {
                      "type": "number"
                    },
                    "utilizationPct": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "roomId",
                    "serviceId",
                    "patientsServed",
                    "busyTimeMs",
                    "utilizationPct"
                  ]
                }
              },
              "events": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "simTimeMs": {
                      "type": "number"
                    },
                    "type": {
                      "type": "string",
                      "enum": [
                        "ARRIVED",
                        "STEP_READY",
                        "QUEUED",
                        "SERVICE_STARTED",
                        "SERVICE_COMPLETED",
                        "PATIENT_COMPLETED"
                      ]
                    },
                    "patientId": {
                      "type": "string"
                    },
                    "serviceId": {
                      "type": "string"
                    },
                    "roomId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "simTimeMs",
                    "type",
                    "patientId"
                  ]
                }
              }
            },
            "required": [
              "algorithm",
              "algorithmLabel",
              "seed",
              "workflow",
              "patientCount",
              "simulationTimeMs",
              "metrics",
              "rooms",
              "events"
            ]
          }
        },
        "required": [
          "success",
          "statusCode",
          "requestId",
          "timestamp",
          "data"
        ]
      },
      "CompareBenchmarkDto": {
        "type": "object",
        "properties": {
          "patientCount": {
            "type": "integer",
            "minimum": 1,
            "exclusiveMinimum": false,
            "maximum": 1000,
            "exclusiveMaximum": false
          },
          "workflow": {
            "type": "string",
            "enum": [
              "INDEPENDENT",
              "SEQUENTIAL",
              "PARTIAL"
            ]
          },
          "seed": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": false,
            "maximum": 2147483647,
            "exclusiveMaximum": false,
            "default": 20261002
          },
          "algorithms": {
            "type": "array",
            "minItems": 2,
            "maxItems": 5,
            "items": {
              "type": "string",
              "enum": [
                "SYSTEM",
                "SHORTEST_QUEUE",
                "ROUND_ROBIN",
                "RANDOM",
                "LEAST_UTILISED"
              ]
            }
          }
        },
        "required": [
          "patientCount",
          "workflow",
          "algorithms"
        ]
      },
      "PaginatedActivityLogEnvelopeDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "statusCode": {
            "type": "integer"
          },
          "requestId": {
            "type": "string"
          },
          "timestamp": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "properties": {
              "items": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "userId": {
                      "type": "string",
                      "nullable": true
                    },
                    "action": {
                      "type": "string"
                    },
                    "entity": {
                      "type": "string"
                    },
                    "entityId": {
                      "type": "string"
                    },
                    "ipAddress": {
                      "type": "string",
                      "nullable": true
                    },
                    "userAgent": {
                      "type": "string",
                      "nullable": true
                    },
                    "metadata": {
                      "nullable": true
                    },
                    "createdAt": {}
                  },
                  "required": [
                    "id",
                    "userId",
                    "action",
                    "entity",
                    "entityId",
                    "ipAddress",
                    "userAgent",
                    "metadata",
                    "createdAt"
                  ]
                }
              },
              "total": {
                "type": "number"
              },
              "page": {
                "type": "number"
              },
              "limit": {
                "type": "number"
              }
            },
            "required": [
              "items",
              "total",
              "page",
              "limit"
            ]
          }
        },
        "required": [
          "success",
          "statusCode",
          "requestId",
          "timestamp",
          "data"
        ]
      },
      "RunBenchmarkDto": {
        "type": "object",
        "properties": {
          "patientCount": {
            "type": "integer",
            "minimum": 1,
            "exclusiveMinimum": false,
            "maximum": 1000,
            "exclusiveMaximum": false
          },
          "workflow": {
            "type": "string",
            "enum": [
              "INDEPENDENT",
              "SEQUENTIAL",
              "PARTIAL"
            ]
          },
          "seed": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": false,
            "maximum": 2147483647,
            "exclusiveMaximum": false,
            "default": 20261002
          },
          "algorithm": {
            "type": "string",
            "enum": [
              "SYSTEM",
              "SHORTEST_QUEUE",
              "ROUND_ROBIN",
              "RANDOM",
              "LEAST_UTILISED"
            ]
          }
        },
        "required": [
          "patientCount",
          "workflow",
          "algorithm"
        ]
      },
      "StaffCurrentSessionEnvelopeDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "statusCode": {
            "type": "integer"
          },
          "requestId": {
            "type": "string"
          },
          "timestamp": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "email": {
                "type": "string"
              },
              "role": {
                "type": "string",
                "enum": [
                  "ADMIN",
                  "DOCTOR",
                  "NURSE",
                  "RECEPTIONIST",
                  "LAB_TECHNICIAN",
                  "PHARMACIST",
                  "CASHIER"
                ],
                "x-enumNames": [
                  "ADMIN",
                  "DOCTOR",
                  "NURSE",
                  "RECEPTIONIST",
                  "LAB_TECHNICIAN",
                  "PHARMACIST",
                  "CASHIER"
                ]
              },
              "status": {
                "type": "string",
                "enum": [
                  "ACTIVE",
                  "INACTIVE",
                  "LOCKED"
                ],
                "x-enumNames": [
                  "ACTIVE",
                  "INACTIVE",
                  "LOCKED"
                ]
              },
              "lastLoginAt": {
                "nullable": true
              },
              "createdAt": {},
              "profile": {
                "type": "object",
                "properties": {
                  "fullName": {
                    "type": "string"
                  },
                  "phone": {
                    "type": "string",
                    "nullable": true
                  },
                  "avatarUrl": {
                    "type": "string",
                    "nullable": true
                  },
                  "gender": {
                    "type": "string",
                    "enum": [
                      "MALE",
                      "FEMALE",
                      "OTHER"
                    ],
                    "x-enumNames": [
                      "MALE",
                      "FEMALE",
                      "OTHER"
                    ],
                    "nullable": true
                  },
                  "birthday": {
                    "nullable": true
                  },
                  "address": {
                    "type": "string",
                    "nullable": true
                  },
                  "description": {
                    "type": "string",
                    "nullable": true
                  }
                },
                "required": [
                  "fullName",
                  "phone",
                  "avatarUrl",
                  "gender",
                  "birthday",
                  "address",
                  "description"
                ],
                "nullable": true
              },
              "effectiveRoles": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "effectivePermissions": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "workspace": {
                "type": "string",
                "enum": [
                  "ADMIN",
                  "DOCTOR"
                ],
                "nullable": true
              }
            },
            "required": [
              "id",
              "email",
              "role",
              "status",
              "lastLoginAt",
              "createdAt",
              "profile",
              "effectiveRoles",
              "effectivePermissions",
              "workspace"
            ]
          }
        },
        "required": [
          "success",
          "statusCode",
          "requestId",
          "timestamp",
          "data"
        ]
      }
    }
  }
}